rename: Typography 컴포넌트 폴더 수정
