+++ src/components/topBar/cancelTopBar/index.tsx
@@ -1,5 +1,6 @@
 import SvgIcons from '@/assets/svgIcons';
-import Typography from '@/components/Typography';
+import Typography from '@/components/typography';
+
 import palette from '@/styles/colors';
 import {StyleSheet, View} from 'react-native';
