+++ src/components/topBar/menuTopBar/index.tsx
@@ -1,5 +1,5 @@
 import SvgIcons from '@/assets/svgIcons';
-import Typography from '@/components/Typography';
+import Typography from '@/components/typography';
 import palette from '@/styles/colors';
 import {View, StyleSheet} from 'react-native';
